+++ ratings/selector/src/commonMain/kotlin/Selector.kt
@@ -4,5 +4,5 @@ import com.soywiz.klock.DateTime
 import dev.inmo.plaguposter.posts.models.PostId
 
 interface Selector {
-    suspend fun take(n: Int = 1, now: DateTime = DateTime.now()): List<PostId>
+    suspend fun take(n: Int = 1, now: DateTime = DateTime.now(), exclude: List<PostId> = emptyList()): List<PostId>
 }