+++ triggers/selector_with_scheduling/src/commonMain/kotlin/PackageInfo.kt
@@ -1,1 +1,0 @@
-package dev.inmo.plaguposter.triggers.selector_with_scheduling