+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 0.1.1
 
+* Update dependencies
+
 ## 0.0.10
 
 ## 0.0.9
 
-* Update depedencies
+* Update dependencies