+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # PlaguPoster
 
+## 0.1.1
+
+* Update dependencies
+* `Triggers`
+  * `SelectorWithTimer`
+    * Opportunity to get schedule of posts using `publishing_autoschedule` command
+
 ## 0.0.10
 
 ## 0.0.9
 
-* Update depedencies
+* Update dependencies